refactor(keyring): address CodeRabbit review — split mod.rs, fix races and silent errors
- Fix 1: propagate serde_json serialisation failure in FileBackend::write_map
  instead of silently returning empty data via unwrap_or_default; add
  KeyringError::Backend variant to carry the message.
- Fix 2: add parking_lot::Mutex to FileBackend covering the full
  read→modify→write cycle in set() and delete() for in-process safety.
- Fix 3: split heavy mod.rs into ops.rs (get/set/delete/migrate/helpers)
  and store.rs (BACKEND/WORKSPACE_DIR OnceLocks, backend selection);
  mod.rs is now re-exports only.
- Fix 4: force_backend_for_test panics loudly if BACKEND is already
  initialised instead of silently ignoring the set() failure.
- Fix 5: rewrite migration tests to call the production migrate_from_file
  logic (via run_migrate helper driving FileBackend directly) and assert
  on MigrationOutcome variants + post-conditions (file removed, value stored).
